+++ protocols/__init__.py
@@ -2,6 +2,8 @@
 
 from protocols.messages import ActionRequest, Decision, Observation
 from protocols.view_filter import build_observation, filter_state_for_seat
+from protocols.player_view_builder import PlayerViewBuilder
+from protocols.player_view_v2 import PLAYER_VIEW_VERSION, PlayerViewV2
 
 __all__ = [
     "ActionRequest",
@@ -9,4 +11,7 @@ __all__ = [
     "Observation",
     "build_observation",
     "filter_state_for_seat",
+    "PLAYER_VIEW_VERSION",
+    "PlayerViewBuilder",
+    "PlayerViewV2",
 ]